download button working and fixes
download files working and fixed a bit of the layout of the files. Fixed long file names to have "...", along with a hover over to show the full name. Made the file names start lower in the screen to not get cut off by the nav bar when viewing images